+++ src/main/kotlin/com/greenkode/trader/event/OrderEvent.kt
@@ -4,16 +4,17 @@ import com.greenkode.trader.domain.EventTypeEnum
 import com.greenkode.trader.domain.OrderAction
 import com.greenkode.trader.domain.OrderType
 import com.greenkode.trader.domain.Symbol
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 
 class OrderEvent(
     val symbol: Symbol,
     val orderType: OrderType,
-    val quantity: Double,
+    val quantity: BigDecimal,
     val action: OrderAction,
     val timestamp: LocalDateTime,
-    val price: Double
+    val price: BigDecimal
 ) : Event {
 
     override val type: EventTypeEnum
