+++ src/main/kotlin/com/greenkode/trader/portfolio/Portfolio.kt
@@ -2,6 +2,7 @@ package com.greenkode.trader.portfolio
 
 import com.greenkode.trader.domain.Symbol
 import com.greenkode.trader.event.Event
+import java.math.BigDecimal
 
 abstract class Portfolio {
 
@@ -13,7 +14,7 @@ abstract class Portfolio {
 
     abstract fun getHistoricalHoldings(): List<Holdings>
 
-    abstract fun getCurrentPositions(): Map<Symbol, Double>
+    abstract fun getCurrentPositions(): Map<Symbol, BigDecimal>
 
-    abstract fun getCurrentHoldings(): Map<Symbol, Double>
+    abstract fun getCurrentHoldings(): Map<Symbol, BigDecimal>
 }
